epoch plot

+++ imdb/run_imdb.py
@@ -228,7 +228,8 @@ def train(model,
                         epochs_vs_performance['epoch'].append(epoch)
                         epochs_vs_performance['step'].append(global_step)
                         epochs_vs_performance['f1_score'].append(
-                            f1_score(labels.tolist(), output.tolist()))
+                            f1_score(labels.astype(int).tolist(),
+                                     output.astype(int).tolist()))
 
                         loss = criterion(output, labels)
                         valid_running_loss += loss.item()
